requirements.txt
django==2.2.1
django_filters
djangorestframework
Markdown